Update to Full Screen Module due to platform upgrade
Boomi have created an action button that creates a secodney selection to move into full screen, this breaks the entire full screen module and the existing is not reusable due to these changes.

There is also not specific selectors that can be leveraged as part of this so the code is now re-duced to leverage the base full on/off process.