Notification close button will change header color to yellow to indicate revision
Floated the notification button to the top right of the notification div
Added mouse-pointer cursor to element to indicate you can click it
When closing the notification it will now set the tab color to yellow to indicate the tab is in "revision mode". Hopefully this will also clarify the lack of a "Test" button as revisions cannot be run in Test mode.