Fixed options inserted to ellipse menu
In the new February 2024 update they changed the gear icon to an ellipse icon. This also changed the element from a <button> to <a>... so to work with both approaches we wait for the Enter Full Screen button to appear in the menu, traverse up the tree to find the list and then insert the UI elements.